requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.36
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic-settings==2.1.0
alembic==1.14.0
# Async database drivers
asyncpg>=0.30.0
aiomysql>=0.3.2
pymysql>=1.1.2
# greenlet is installed automatically by SQLAlchemy, but we can specify a compatible version
# Using latest version that supports Python 3.12
greenlet>=3.1.0
python-dotenv==1.0.0
email-validator==2.1.0
bcrypt==4.0.1
reportlab==4.2.5
openpyxl==3.1.5
phonenumbers==8.13.31
aiohttp==3.9.1
httpx==0.26.0
cryptography==41.0.7
sentry-sdk[fastapi]==2.15.0
redis==5.0.1
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-cov==5.0.0
aiosqlite==0.19.0
Pillow>=10.4.0,<11.0.0
pywebpush==1.14.0
twilio==9.3.0
pyotp==2.9.0
qrcode[pil]==7.4.2
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
# AI Integration Libraries
openai>=1.12.0
anthropic>=0.18.0
google-generativeai>=0.3.0
openai[azure]>=1.12.0
# Voice Transcription Libraries
SpeechRecognition==3.10.0
pydub==0.25.1
# AWS Comprehend Medical
boto3>=1.34.0
botocore>=1.34.0

